if one is going to use size_t all over the place, one should be prepared to use size_t *all over the place*.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@17148 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/other/ftgl/src/FTCharmap.h
@@ -121,7 +121,7 @@ class FTCharmap
          *                       character code.
          */
         void InsertIndex(const unsigned int characterCode,
-                         const unsigned int containerIndex);
+                         const size_t containerIndex);
 
         /**
          * Queries for errors.
